refactor(ramp): remove RampActivityDetails component and related files
