+++ app/components/Views/ActivityDetails/hooks/useActivityDetailsItem.ts
@@ -5,6 +5,7 @@ import type { ActivityListItem } from '../../../../util/activity-adapters';
 import { selectNonEvmTransactionsForSelectedAccountGroup } from '../../../../selectors/multichain/multichain';
 /* eslint-disable import-x/no-restricted-paths -- TODO(ADR-0020): reuses the activity list's data sources; route-isolation backlog */
 import { useLocalActivityItems } from '../../ActivityList/hooks/useLocalActivityItems';
+import { useRampActivityItems } from '../../ActivityList/hooks/useRampActivityItems';
 import { useTransactionsQuery } from '../../ActivityList/useTransactionsQuery';
 import { mapNonEvmTransactions } from '../../ActivityList/helpers/transformations';
 /* eslint-enable import-x/no-restricted-paths */
@@ -47,7 +48,8 @@ function buildItemsByIdentifier(
   for (const item of items) {
     const identifier =
       item.raw?.type === 'perpsTransaction' ||
-      item.raw?.type === 'predictActivity'
+      item.raw?.type === 'predictActivity' ||
+      item.raw?.type === 'rampOrder'
         ? item.raw.data.id
         : undefined;
     const normalizedIdentifier = identifier?.toLowerCase();
@@ -77,6 +79,7 @@ export function useActivityDetailsItem(
   preloadedItem?: ActivityListItem,
 ): ActivityListItem | undefined {
   const localActivityItems = useLocalActivityItems();
+  const rampActivityItems = useRampActivityItems();
   const { data: evmTransactions } = useTransactionsQuery();
   const nonEvmState = useSelector(
     selectNonEvmTransactionsForSelectedAccountGroup,
@@ -111,6 +114,10 @@ export function useActivityDetailsItem(
       ),
     [preloadedItem, chainId],
   );
+  const rampByIdentifier = useMemo(
+    () => buildItemsByIdentifier(filterByChain(rampActivityItems, chainId)),
+    [rampActivityItems, chainId],
+  );
 
   return useMemo(() => {
     const id = txIdentifier?.toLowerCase();
@@ -122,9 +129,14 @@ export function useActivityDetailsItem(
     const apiItem = apiByHash.get(id);
     const nonEvmItem = nonEvmByHash.get(id);
     const preloadedResolvedItem = preloadedByIdentifier.get(id);
+    const rampItem = rampByIdentifier.get(id);
 
     if (preloadedResolvedItem) {
       return preloadedResolvedItem;
+    }
+
+    if (rampItem) {
+      return rampItem;
     }
 
     if (localItem) {
@@ -143,5 +155,6 @@ export function useActivityDetailsItem(
     apiByHash,
     nonEvmByHash,
     preloadedByIdentifier,
+    rampByIdentifier,
   ]);
 }
